+++ app/src/main/java/labs/zero_one/roundo/MainActivity.kt
@@ -3,15 +3,13 @@ package labs.zero_one.roundo
 import android.app.Activity
 import android.app.AlertDialog
 import android.content.Intent
-import android.content.pm.PackageManager
-import android.location.LocationManager
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
 import android.support.v7.app.AppCompatActivity
 import android.view.Menu
 import android.view.MenuItem
 import com.minemap.minemapsdk.MinemapAccountManager
 import com.minemap.minemapsdk.camera.CameraPosition
+import com.minemap.minemapsdk.geometry.LatLng
 import com.minemap.minemapsdk.maps.MineMap
 import com.minemap.minemapsdk.maps.OnMapReadyCallback
 import kotlinx.android.synthetic.main.activity_main.*
@@ -21,6 +19,7 @@ import kotlinx.android.synthetic.main.content_main.*
  * 主页面 Activity
  *
  * 属性列表
+ * [mineMap]
  *
  * 子类列表
  * [MainMenu]
@@ -33,13 +32,16 @@ import kotlinx.android.synthetic.main.content_main.*
  * [onActivityResult]
  *
  * 自定义方法列表
+ * [initMap]
  *
  * @author lucka-me
  * @since 0.1
+ *
+ * @property [mineMap] 地图控制器
  */
 class MainActivity : AppCompatActivity() {
 
-    //private lateinit var mineMap: MineMap
+    private lateinit var mineMap: MineMap
 
     /**
      * 主菜单项
@@ -47,14 +49,17 @@ class MainActivity : AppCompatActivity() {
      * @param [index] 菜单项位置
      * @param [id] 菜单项资源 ID
      *
-     * @property [StartStop] 开始/停止
-     * @property [Preference] 设置
-     *
      * @author lucka
      * @since 0.1
      */
     private enum class MainMenu(val index: Int, val id: Int) {
+        /**
+         * 开始/停止
+         */
         StartStop(0, R.id.menu_main_start_stop),
+        /**
+         * 设置
+         */
         Preference(1, R.id.menu_main_preference)
     }
 
@@ -63,12 +68,13 @@ class MainActivity : AppCompatActivity() {
      *
      * @param [code] 请求代码
      *
-     * @property [Setup] 准备任务
-     *
      * @author lucka-me
      * @since 0.1.3
      */
     private enum class ActivityRequest(val code: Int) {
+        /**
+         * 准备 Activity
+         */
         Setup(1)
     }
 
@@ -78,9 +84,12 @@ class MainActivity : AppCompatActivity() {
         setSupportActionBar(mainToolbar)
 
         // Setup Map
-        /*MinemapAccountManager.getInstance(applicationContext, "9282e20b60d248c5b0a2048ff39b625d", "4787")
-        initMap(savedInstanceState)*/
-
+        MinemapAccountManager.getInstance(
+            applicationContext,
+            getString(R.string.minemap_token),
+            "4807"
+        )
+        initMap(savedInstanceState)
 
     }
 
@@ -130,15 +139,22 @@ class MainActivity : AppCompatActivity() {
 
         }
     }
-/*
+
+    /**
+     * 初始化地图
+     *
+     * @param [savedInstanceState] 初始化所需参数
+     *
+     * @author lucka-me
+     * @since 0.1.4
+     */
     private fun initMap(savedInstanceState: Bundle?) {
         mapView.onCreate(savedInstanceState)
-        mapView.setStyleUrl("http://minedata.cn/service/solu/style/id/4787")
         mapView.getMapAsync(OnMapReadyCallback { newMap: MineMap? ->
             if (newMap == null) return@OnMapReadyCallback
             mineMap = newMap
             mineMap.uiSettings.isCompassEnabled = true
-            mineMap.cameraPosition = CameraPosition.DEFAULT
+            mineMap.cameraPosition = CameraPosition.Builder().target(LatLng(34.2651799, 108.9435278)).zoom(10.0).build()
         })
-    }*/
+    }
 }